backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic-settings==2.6.1
sqlalchemy==2.0.36
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
